upd config and create populate
