+++ lantern/flask/flask_SQLalchemy/practice/app.py
@@ -2,16 +2,7 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy_utils import create_database, drop_database, database_exists
 
-
-
-class Config:
-    PG_USER = "cursor"
-    PG_PASSWORD = "very_secret_password"
-    PG_HOST = "localhost"
-    PG_PORT = 5432
-    DB_NAME = "cursor_sqlalchemy_db"
-    SQLALCHEMY_DATABASE_URI = f"postgresql://{PG_USER}:{PG_PASSWORD}@{PG_HOST}:{PG_PORT}/{DB_NAME}"
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
+from config import Config
 
 
 db = SQLAlchemy()
@@ -24,6 +15,6 @@ with app.app_context():
         db.create_all()
         print('Database exists')
     else:
-        print("Database does not exists")
-
-
+        print(f"Database does not exists {db.engine.url}")
+        create_database(db.engine.url)
+        print('Data base created')
