+++ lantern/flask/flask_SQLalchemy/practice/config.py
@@ -1,0 +1,12 @@
+
+class Config:
+    PG_USER = "cursor"
+    PG_PASSWORD = "very_secret_password"
+    PG_HOST = "localhost"
+    PG_PORT = 5432
+    DB_NAME = "cursor_sqlalchemy_db"
+    SQLALCHEMY_DATABASE_URI = f"postgresql://{PG_USER}:{PG_PASSWORD}@{PG_HOST}:{PG_PORT}/{DB_NAME}"
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+
+
